+++ problems/inheritance-problems/pedigrees/write_pedigree_choice.py
@@ -11,9 +11,9 @@ import random
 # Import local modules from the project
 # Provides custom functions, such as question formatting and other utilities
 import bptools
-import pedigree_html_lib
-import pedigree_code_lib
-import pedigree_code_templates
+import pedigree_lib.html_output
+import pedigree_lib.code_definitions
+import pedigree_lib.code_templates
 
 
 #=======================
@@ -24,42 +24,42 @@ def multipleChoiceQuestionSet(start_num: int):
 	+"Which one of the following patterns of inheritance is most likely demonstrated in the above pedigree inheritance?</p> "
 	)
 	N = start_num - 1
-	for ad in pedigree_code_templates.autosomal_dominant:
+	for ad in pedigree_lib.code_templates.autosomal_dominant:
 		if random.random() < 0.5:
-			ad = pedigree_code_lib.mirror_pedigree(ad)
-		adc = pedigree_html_lib.translateCode(ad)
+			ad = pedigree_lib.code_definitions.mirror_pedigree(ad)
+		adc = pedigree_lib.html_output.translateCode(ad)
 		answer = 'autosomal dominant'
 		N += 1
 		bb_output_format = bptools.formatBB_MC_Question(N, adc+question_text, choices_list, answer)
 		bb_output_format_list.append(bb_output_format)
-	for ar in pedigree_code_templates.autosomal_recessive:			
+	for ar in pedigree_lib.code_templates.autosomal_recessive:			
 		if random.random() < 0.5:
-			ar = pedigree_code_lib.mirror_pedigree(ar)
-		arc = pedigree_html_lib.translateCode(ar)
+			ar = pedigree_lib.code_definitions.mirror_pedigree(ar)
+		arc = pedigree_lib.html_output.translateCode(ar)
 		answer = 'autosomal recessive'
 		N += 1
 		bb_output_format = bptools.formatBB_MC_Question(N, arc+question_text, choices_list, answer)
 		bb_output_format_list.append(bb_output_format)
-	for xd in pedigree_code_templates.x_linked_dominant:
+	for xd in pedigree_lib.code_templates.x_linked_dominant:
 		if random.random() < 0.5:
-			xd = pedigree_code_lib.mirror_pedigree(xd)
-		xdc = pedigree_html_lib.translateCode(xd)
+			xd = pedigree_lib.code_definitions.mirror_pedigree(xd)
+		xdc = pedigree_lib.html_output.translateCode(xd)
 		answer = 'x-linked dominant'
 		N += 1
 		bb_output_format = bptools.formatBB_MC_Question(N, xdc+question_text, choices_list, answer)
 		bb_output_format_list.append(bb_output_format)
-	for xr in pedigree_code_templates.x_linked_recessive:
+	for xr in pedigree_lib.code_templates.x_linked_recessive:
 		if random.random() < 0.5:
-			xr = pedigree_code_lib.mirror_pedigree(xr)
-		xrc = pedigree_html_lib.translateCode(xr)
+			xr = pedigree_lib.code_definitions.mirror_pedigree(xr)
+		xrc = pedigree_lib.html_output.translateCode(xr)
 		answer = 'x-linked recessive'
 		N += 1
 		bb_output_format = bptools.formatBB_MC_Question(N, xrc+question_text, choices_list, answer)
 		bb_output_format_list.append(bb_output_format)
-	for yl in pedigree_code_templates.y_linked:
+	for yl in pedigree_lib.code_templates.y_linked:
 		if random.random() < 0.5:
-			yl = pedigree_code_lib.mirror_pedigree(yl)
-		ylc = pedigree_html_lib.translateCode(yl)
+			yl = pedigree_lib.code_definitions.mirror_pedigree(yl)
+		ylc = pedigree_lib.html_output.translateCode(yl)
 		answer = 'y-linked'
 		N += 1
 		bb_output_format = bptools.formatBB_MC_Question(N, ylc+question_text, choices_list, answer)